Restructure file system to make components more modular. Import a normalize.css template into index.css
