+++ friends/src/components/FriendComponents/Friend.js
